optimizing video rendering

+++ PainterlyRendering/SplineStroke.cpp
@@ -9,8 +9,8 @@ void SplineStroke::cubic_b_spline(Point2d &p, double t) const
 	if (npoints == 0) return;
 	if (t > 1) t = 1;
 
-	index = (int)(t * (npoints - 3.0));
-	tt = (double) (t - (double)index / ((double)npoints - 3.0)) * (npoints - 3.0);
+	index = t * (npoints - 3.0);
+	tt = (t - (double)index / ((double)npoints - 3.0)) * (npoints - 3.0);
 
 	if (index < 0) index = 0;
 	if (index > npoints - 4) index = npoints - 4;
